Move IShaderPath
