+++ packages/dev/core/src/Materials/effect.ts
@@ -20,7 +20,45 @@ import type { InternalTexture } from "../Materials/Textures/internalTexture";
 import type { ThinTexture } from "../Materials/Textures/thinTexture";
 import type { RenderTargetTexture } from "../Materials/Textures/renderTargetTexture";
 import type { PostProcess } from "../PostProcesses/postProcess";
-import type { IShaderPath } from "./shaderMaterial";
+
+/**
+ * Defines the route to the shader code. The priority is as follows:
+ *  * object: `{ vertexSource: "vertex shader code string", fragmentSource: "fragment shader code string" }` for directly passing the shader code
+ *  * object: `{ vertexElement: "vertexShaderCode", fragmentElement: "fragmentShaderCode" }`, used with shader code in script tags
+ *  * object: `{ vertex: "custom", fragment: "custom" }`, used with `Effect.ShadersStore["customVertexShader"]` and `Effect.ShadersStore["customFragmentShader"]`
+ *  * string: `"./COMMON_NAME"`, used with external files COMMON_NAME.vertex.fx and COMMON_NAME.fragment.fx in index.html folder.
+ */
+export type IShaderPath =
+    | {
+          /**
+           * Directly pass the shader code
+           */
+          vertexSource?: string;
+          /**
+           * Directly pass the shader code
+           */
+          fragmentSource?: string;
+
+          /**
+           * Used with Effect.ShadersStore. If the `vertex` is set to `"custom`, then
+           * Babylon.js will read from Effect.ShadersStore["customVertexShader"]
+           */
+          vertex?: string;
+          /**
+           * Used with Effect.ShadersStore. If the `fragment` is set to `"custom`, then
+           * Babylon.js will read from Effect.ShadersStore["customFragmentShader"]
+           */
+          fragment?: string;
+          /**
+           * Used with shader code in script tags
+           */
+          vertexElement?: string;
+          /**
+           * Used with shader code in script tags
+           */
+          fragmentElement?: string;
+      }
+    | string;
 
 /**
  * Options to be used when creating an effect.
